+++ multiCheckbox/index.html
@@ -82,7 +82,7 @@
       </ul>
     </div>
   </div>
-  <script src="app.js"></script>
+  <script src="app.js" type="module"></script>
 </body>
 
 </html>
